requirements.txt
fastapi
uvicorn[standard]
llama-index-readers-file
llama-index-llms-openai
llama-index-embeddings-openai
python-dotenv